# By Jose M. Olea (2) and Alger Ortín Castellví (1)
# Via Jose M. Olea (2) and Alger Ortín Castellví (1)
* 'feature/metagame/shop_redesign' of bcn-mb-git.ubisoft.org:dragon/client:
  Fix some issues produced by the merge.
  Rotational pill offers introduced. Still WIP.
  Shop: Development - Unified currency packs as offer packs. Shop: Content - Refactored shopPacksDefinitions table so it can be directly parsed as Offer Packs. Shop HC Pill: Development - New type of pill specialized for HC packs. Shop HC Item Slot: Development - New type of item slot customized for HC pills. HC and PC Item Previews: Development - New type of item previews for the shop.


Former-commit-id: 672a5dc2abac098d0435bfd663f3116e5aa978b3

+++ Assets/Game/Model/Offers/HappyHour.cs
@@ -210,11 +210,11 @@ public class HappyHour {
 	/// Apply the extra amount formula to the given base amount using this Happy Hour's current data.
 	/// </summary>
 	/// <returns>The total final amount.</returns>
-	public int ApplyHappyHourExtra(int _amount) {
+	public long ApplyHappyHourExtra(long _amount) {
 		// Only if active
 		if(!IsActive()) return _amount;
 		
 		// Apply the extra gems factor
-		return Mathf.RoundToInt((_amount) * (1 + extraGemsFactor));
+		return (long)Mathf.RoundToInt((_amount) * (1 + extraGemsFactor));
 	}
 }
